Gestionnaire de tâche normalement fini. Maintenant, le consommateur.

+++ requete_reponse.h
@@ -6,7 +6,7 @@
 typedef struct {
 	pid_t pid_processus; /* permet de déterminer si c'est une connexion ou un calcul (comparaison) */
 	int connexion; /* 1 : connexion, 0 : deconnexion, 3 : shutdown */
-	double reponse_calcul;
+	unsigned long int reponse_calcul;
 } reponse_t;
 
 #endif /*MINI_PROJET_2_REQUETE_REPONSE_H*/
